requirements.txt
python-telegram-bot>=21.0
Pillow>=10.2.0
jdatetime==4.1.1
python-dotenv==1.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
